requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
tqdm>=4.66.0
openai>=1.30.0
litellm>=1.40.0
langextract>=0.1.0
pdfplumber>=0.10.0
pypdf>=4.0.0
pyspark>=3.5.0
pyarrow>=14.0.0
findspark>=2.0.1
colorama>=0.4.6
rich>=13.7.0
fastapi>=0.110.0
uvicorn>=0.29.0
databricks-sdk
